+++ src/MediaLibraryServiceProvider.php
@@ -3,6 +3,9 @@
 namespace Spatie\MediaLibrary;
 
 use Illuminate\Support\ServiceProvider;
+use Laravel\Lumen\Application as LumenApplication;
+use Illuminate\Foundation\Application as LaravelApplication;
+use Spatie\MediaLibrary\Commands\CleanCommand;
 use Spatie\MediaLibrary\Commands\ClearCommand;
 use Spatie\MediaLibrary\Commands\RegenerateCommand;
 
@@ -20,22 +23,25 @@ class MediaLibraryServiceProvider extends ServiceProvider
      */
     public function boot()
     {
+        if ($this->app instanceof LaravelApplication) {
+            $this->publishes([
+                __DIR__.'/../config/laravel-medialibrary.php' => config_path('laravel-medialibrary.php'),
+            ], 'config');
+
+            if (!class_exists('CreateMediaTable')) {
+                // Publish the migration
+                $timestamp = date('Y_m_d_His', time());
+
+                $this->publishes([
+                    __DIR__.'/../database/migrations/create_media_table.php.stub' => database_path('migrations/'.$timestamp.'_create_media_table.php'),
+                  ], 'migrations');
+            }
+        } elseif ($this->app instanceof LumenApplication) {
+            $this->app->configure('laravel-medialibrary');
+        }
+
         $mediaClass = config('laravel-medialibrary.media_model');
         $mediaClass::observe(new MediaObserver());
-
-        $this->publishes([
-            __DIR__.'/../config/laravel-medialibrary.php' => config_path('laravel-medialibrary.php'),
-        ], 'config');
-
-        if (!class_exists('CreateMediaTable')) {
-
-            // Publish the migration
-            $timestamp = date('Y_m_d_His', time());
-
-            $this->publishes([
-                __DIR__.'/../database/migrations/create_media_table.php.stub' => database_path('migrations/'.$timestamp.'_create_media_table.php'),
-            ], 'migrations');
-        }
     }
 
     /**
@@ -49,10 +55,12 @@ class MediaLibraryServiceProvider extends ServiceProvider
 
         $this->app->bind('command.medialibrary:regenerate', RegenerateCommand::class);
         $this->app->bind('command.medialibrary:clear', ClearCommand::class);
+        $this->app->bind('command.medialibrary:clean', CleanCommand::class);
 
         $this->commands([
             'command.medialibrary:regenerate',
             'command.medialibrary:clear',
+            'command.medialibrary:clean',
         ]);
     }
 }